pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Ole-Jakob Olsen", email = "ole.jakob.olsen@ngi.no" }
]
requires-python = ">=3.13"
dependencies = [
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.121.1",
    "htpy>=25.10.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
]

[project.scripts]
dev = "scripts.dev:main"
build = "scripts.build:main"
main = "main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
